+++ backend/src/main/java/com/nova/backend/user/service/MyPageService.java
@@ -4,6 +4,6 @@ import com.nova.backend.user.dto.MyPageRequestDTO;
 import com.nova.backend.user.dto.MyPageResponseDTO;
 
 public interface MyPageService {
-    MyPageResponseDTO findByUserId(int userId);
+    MyPageResponseDTO findByUserId(Long userId);
     void updateMyPage(MyPageRequestDTO myPageRequestDTO);
 }
